backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Meme Generator Backend"
authors = ["Jon-Mark Hampson"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.38.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "google>=3.0.0",
    "google-genai>=1.43.0",
    "isort>=7.0.0",
    "logfire[fastapi,httpx,psycopg2,sqlalchemy]>=3.19.0",
    "openai>=1.86.0",
    "passlib>=1.7.4",
    "pillow>=11.2.1",
    "psycopg2>=2.9.10",
    "pydantic-ai>=0.2.16",
    "pydantic-ai-slim[duckduckgo,google]>=0.2.16",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "supabase>=2.15.2",
    "uvicorn>=0.34.2",
]
[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
line_length = 79
